requirements.txt
fastapi
uvicorn
requests
apscheduler
pydantic
secure-smtplib
python-dotenv